Refactor drawEllipseByBoundingBox
