+++ modules/core/src/lib/layers/editable-geojson-layer.js
@@ -3,7 +3,6 @@
 
 import { GeoJsonLayer, ScatterplotLayer, IconLayer } from 'deck.gl';
 import bboxPolygon from '@turf/bbox-polygon';
-import circle from '@turf/circle';
 import ellipse from '@turf/ellipse';
 import distance from '@turf/distance';
 import center from '@turf/center';
@@ -356,7 +355,6 @@ export default class EditableGeoJsonLayer extends EditableLayer {
 
     // if (
     //   this.props.mode === 'drawRectangleUsing3Points' ||
-    //   this.props.mode === 'drawEllipseByBoundingBox' ||
     //   this.props.mode === 'drawEllipseUsing3Points'
     // ) {
     //   } else if (selectedFeatures.length === 1) {
@@ -365,14 +363,6 @@ export default class EditableGeoJsonLayer extends EditableLayer {
     //       this.props.mode === 'drawEllipseUsing3Points'
     //     ) {
     //       this.handleDrawRectangleUsing3Points(
-    //         selectedFeatures[0],
-    //         selectedFeatureIndexes[0],
-    //         groundCoords,
-    //         picks
-    //       );
-    //     }
-    //     if (this.props.mode === 'drawEllipseByBoundingBox') {
-    //       this.handleDrawEllipseByBoundingBox(
     //         selectedFeatures[0],
     //         selectedFeatureIndexes[0],
     //         groundCoords,
@@ -549,28 +539,6 @@ export default class EditableGeoJsonLayer extends EditableLayer {
             coordinates: [startPosition, endPosition]
           }
         };
-      } else if (mode === 'drawEllipseByBoundingBox') {
-        const corner1 = ((selectedFeature.geometry.coordinates: any): Array<number>);
-        const corner2 = groundCoords || corner1;
-
-        const minX = Math.min(corner1[0], corner2[0]);
-        const minY = Math.min(corner1[1], corner2[1]);
-        const maxX = Math.max(corner1[0], corner2[0]);
-        const maxY = Math.max(corner1[1], corner2[1]);
-
-        const polygonPoints = bboxPolygon([minX, minY, maxX, maxY]).geometry.coordinates[0];
-        const ellipseCenter = center(fc([point(corner1), point(corner2)]));
-
-        const xSemiAxis = Math.max(
-          distance(point(polygonPoints[0]), point(polygonPoints[1])),
-          0.001
-        );
-        const ySemiAxis = Math.max(
-          distance(point(polygonPoints[0]), point(polygonPoints[3])),
-          0.001
-        );
-
-        drawFeature = ellipse(ellipseCenter, xSemiAxis, ySemiAxis);
       }
     } else if (selectedFeature.geometry.type === 'LineString') {
       const positionOfLineString = this.props.drawAtFront
@@ -783,43 +751,6 @@ export default class EditableGeoJsonLayer extends EditableLayer {
   }
 
   handleDrawRectangle(
-    selectedFeature: Feature,
-    featureIndex: number,
-    groundCoords: Position,
-    picks: Object[]
-  ) {
-    let featureCollection = this.state.editableFeatureCollection;
-    let positionIndexes;
-
-    let updatedMode = this.props.mode;
-
-    if (selectedFeature.geometry.type === 'Point') {
-      positionIndexes = null;
-      featureCollection = featureCollection.replaceGeometry(
-        featureIndex,
-        // $FlowFixMe: it's ok, I know drawFeature will be there
-        this.state.drawFeature.geometry
-      );
-      updatedMode = 'modify';
-    } else {
-      console.warn(`Unsupported geometry type: ${selectedFeature.geometry.type}`); // eslint-disable-line
-      return;
-    }
-
-    const updatedData = featureCollection.getFeatureCollection();
-
-    this.props.onEdit({
-      updatedData,
-      updatedMode,
-      updatedSelectedFeatureIndexes: this.props.selectedFeatureIndexes,
-      editType: 'addPosition',
-      featureIndex,
-      positionIndexes,
-      position: groundCoords
-    });
-  }
-
-  handleDrawEllipseByBoundingBox(
     selectedFeature: Feature,
     featureIndex: number,
     groundCoords: Position,
